refactor: convert to more typical project structure with sde.config.js file
