--- examples/house-game/.eslintrc.cjs
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: ['../../.eslintrc-svelte-common.cjs']
-}
