+++ examples/house-game/.gitignore
@@ -1,1 +1,1 @@
-public
+sde-prep
