+++ examples/house-game/README.md
@@ -1,0 +1,7 @@
+# house-game
+
+TODO
+
+## License
+
+SDEverywhere is distributed under the MIT license. See `LICENSE` for more details.
