+++ examples/house-game/model/houses.dat
@@ -1,0 +1,2 @@
+planning data
+0	0
