+++ examples/house-game/package.json
@@ -1,48 +1,20 @@
 {
   "name": "@sdeverywhere/house-game",
-  "private": true,
   "version": "1.0.0",
+  "private": true,
   "type": "module",
   "scripts": {
-    "clean": "rm -rf public",
-    "lint:ts": "eslint src --max-warnings 0",
-    "lint:svelte": "svelte-check --fail-on-warnings --output human --compiler-warnings \"a11y-click-events-have-key-events:ignore,a11y-no-noninteractive-tabindex:ignore\"",
-    "lint": "run-s lint:ts lint:svelte",
-    "prettier:check": "prettier --check .",
-    "prettier:fix": "prettier --write .",
-    "precommit": "../../scripts/precommit",
-    "test": "echo No tests yet",
-    "type-check": "tsc --noEmit -p tsconfig-build.json",
-    "build": "vite build",
-    "dev": "vite",
-    "ci:build": "run-s clean lint prettier:check type-check build"
+    "dev": "sde dev",
+    "build": "sde bundle"
   },
+  "workspaces": [
+    "packages/app"
+  ],
   "dependencies": {
-    "@sdeverywhere/compile": "workspace:*",
-    "@sdeverywhere/runtime": "workspace:*",
-    "chart.js": "^2.9.4"
-  },
-  "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^2.4.5",
-    "@types/chart.js": "^2.9.34",
-    "postcss": "^8.4.31",
-    "pug": "^3.0.1",
-    "sass": "^1.34.1",
-    "svelte": "^3.59.2",
-    "svelte-check": "^3.5.1",
-    "svelte-preprocess": "^5.0.4",
-    "vite": "^4.4.9",
-    "vite-plugin-node-polyfills": "^0.21.0"
-  },
-  "author": "Climate Interactive",
-  "license": "MIT",
-  "homepage": "https://sdeverywhere.org",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/climateinteractive/SDEverywhere.git",
-    "directory": "examples/playground"
-  },
-  "bugs": {
-    "url": "https://github.com/climateinteractive/SDEverywhere/issues"
+    "@sdeverywhere/build": "^0.3.4",
+    "@sdeverywhere/cli": "^0.7.23",
+    "@sdeverywhere/plugin-config": "^0.2.4",
+    "@sdeverywhere/plugin-vite": "^0.1.8",
+    "@sdeverywhere/plugin-worker": "^0.2.3"
   }
 }
