+++ examples/house-game/packages/app/.eslintignore
