+++ examples/house-game/packages/app/.eslintrc.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: ['../../../.eslintrc-svelte-common.cjs']
+}
