+++ examples/house-game/packages/app/.gitignore
@@ -1,0 +1,2 @@
+src/model/generated
+public
