+++ examples/house-game/packages/app/.prettierignore
