+++ examples/house-game/packages/app/index.html
