+++ examples/house-game/packages/app/src/app.svelte
@@ -3,23 +3,24 @@
 
 import './global.css'
 
+import Assumptions from './components/assumptions/assumptions.svelte'
 import Graph from './components/graph/graph.svelte'
 
 import type { AppViewModel } from './app-vm'
 
-export let appViewModel: AppViewModel
-const busy = appViewModel.busy
-const message = appViewModel.message
-const assumptionRows = appViewModel.assumptionRows
-const currentTime = appViewModel.currentTime
-const currentValue = appViewModel.writableCurrentValue
+export let viewModel: AppViewModel
+const busy = viewModel.busy
+const message = viewModel.message
+const assumptionsViewModel = viewModel.assumptions
+const currentTime = viewModel.currentTime
+const currentValue = viewModel.writableCurrentValue
 
 function onContinue() {
-  appViewModel.nextStep()
+  viewModel.nextStep()
 }
 
 function onReset() {
-  appViewModel.reset()
+  viewModel.reset()
 }
 
 </script>
@@ -45,22 +46,14 @@ function onReset() {
     </div>
 
     <div class="text-container assumptions">
-      <div class="assumptions-title">Assumptions</div>
-      <div class="assumption-rows">
-        {#each $assumptionRows as row (row.label)}
-          <div class="assumption-row">
-            <div class="assumption-name">{row.label}</div>
-            <div class="assumption-value">{@html row.value}</div>
-          </div>
-        {/each}
-      </div>
+      <Assumptions viewModel={$assumptionsViewModel} />
     </div>
   </div>
 
   <div class="right-container">
     <div class="column">
       <div class="graph-container">
-        <Graph viewModel={appViewModel.supplyGraphViewModel} width={500} height={400}/>
+        <Graph viewModel={viewModel.supplyGraphViewModel} width={500} height={400}/>
       </div>
     </div>
   </div>
@@ -142,25 +135,6 @@ button
   height: 180px
   gap: 4px
   color: #777
-
-.assumptions-title
-  font-weight: 700
-  margin-bottom: 12px
-
-.assumption-rows
-  display: flex
-  flex-direction: column
-  width: 100%
-
-.assumption-row
-  display: flex
-  flex-direction: row
-  width: 100%
-  justify-content: space-between
-  font-size: 14px
-  line-height: 1.4
-  &:nth-child(2)
-    margin-bottom: 12px
 
 .right-container
   display: flex
