+++ examples/house-game/packages/app/src/components/graph/graph-vm.ts
