+++ examples/house-game/packages/app/src/global.css
