+++ examples/house-game/packages/app/svelte.config.js
