--- examples/house-game/src/index.ts
@@ -1,12 +1,0 @@
-import { createAppViewModel } from './app-vm'
-import App from './app.svelte'
-
-const appViewModel = await createAppViewModel()
-const app = new App({
-  target: document.body,
-  props: {
-    appViewModel
-  }
-})
-
-export default app
