--- examples/house-game/src/types.d.ts
@@ -1,2 +1,0 @@
-// XXX: Workaround for lack of type declarations for the `compile` package
-declare module '@sdeverywhere/compile'
