--- examples/house-game/tsconfig-base.json
@@ -1,16 +1,0 @@
-// This contains the TypeScript configuration that is shared between
-// testing (`tsconfig-test.json`) and production builds (`tsconfig-build.json`).
-{
-  "extends": "../../tsconfig-common.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "allowJs": true,
-    "resolveJsonModule": true,
-    "module": "esnext",
-    "target": "esnext",
-    "noImplicitAny": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "types": ["svelte", "vite/client"]
-  }
-}
