--- examples/house-game/tsconfig-test.json
@@ -1,5 +1,0 @@
-// This contains the TypeScript configuration for testing.
-{
-  "extends": "./tsconfig-base.json",
-  "include": ["src/**/*"]
-}
