+++ pnpm-workspace.yaml
@@ -1,6 +1,7 @@
 packages:
   - packages/*
-  - examples/*
+  - examples/**
   - '!examples/template-*'
+  - '!examples/template-*/**'
   - models
   - tests/**
